requirements.txt
torch>=2.4.0
torchvision>=0.19.0
runpod>=1.7.0
Pillow>=10.0.0
accelerate>=0.33.0
transformers>=4.44.0
sentencepiece>=0.2.0
protobuf>=4.25.0
safetensors>=0.4.0
huggingface_hub>=0.25.0
peft>=0.13.0
